backend/requirements.txt
flask
flask-cors
google-genai
gspread
google-auth
python-dotenv
requests
gunicorn
